Adding new generics.

+++ Source/TreasureGuide.DB/dbo/Tables/Teams/TeamGenericSlots.sql
@@ -4,7 +4,7 @@
     [Position] TINYINT NOT NULL,
     [Type] SMALLINT NOT NULL CONSTRAINT [DF_dbo.TeamGenericSlots_Type] DEFAULT 0,
     [Class] SMALLINT NOT NULL CONSTRAINT [DF_dbo.TeamGenericSlots_Class] DEFAULT 0,
-    [Role] SMALLINT NOT NULL CONSTRAINT [DF_dbo.TeamGenericSlots_Role] DEFAULT 0,
+    [Role] INT NOT NULL CONSTRAINT [DF_dbo.TeamGenericSlots_Role] DEFAULT 0,
     [Sub] BIT NOT NULL CONSTRAINT [DF_dbo.TeamGenericSlots_Sub] DEFAULT 0,
     CONSTRAINT [PK_dbo.TeamGenericSlots] PRIMARY KEY CLUSTERED ([TeamId] ASC, [Position] ASC, [Type] ASC, [Class] ASC, [Role] ASC),
     CONSTRAINT [CK_dbo.TeamGenericSlots_Position] CHECK ([Position] >= 0 AND [Position] < 6),
